requirements.txt
# env and utility
python-dotenv
requests

# Logging
loguru

# Metrics and benchamrking
psutil 
soundfile 
numpy 
pandas

# TTS
gTTS

# STT
openai-whisper 
faster-whisper 

# LLMS
langchain-groq
langchain-ollama

# Rag
sentence_transformers
faiss-cpu

# Langchain
langchain
langchain-core
langchain-huggingface
langchain-community

# LangGraph


# UI
streamlit
streamlit-audiorecorder
